+++ KeireClient/Include/KeireClient/Editor/ScenePlayChangesPanel.h
@@ -16,10 +16,21 @@ namespace KeireEditor
     class ScenePlayChangesPanel final
     {
       public:
-        void Open() noexcept { m_OpenRequested = true; }
+        void Open() noexcept
+        {
+            m_ReviewPending = true;
+            m_OpenRequested = true;
+        }
+        void Close() noexcept
+        {
+            m_ReviewPending = false;
+            m_OpenRequested = false;
+        }
+        [[nodiscard]] bool Pending() const noexcept { return m_ReviewPending; }
         [[nodiscard]] ScenePlayDecision Draw(Keire::UiFrame& ui, ScenePlayChangeSet& changes);
 
       private:
         bool m_OpenRequested = false;
+        bool m_ReviewPending = false;
     };
 } // namespace KeireEditor